Server/app/utils/token_required.py
from flask import request, jsonify
import jwt
from functools import wraps

SECRET_KEY = 'your_secret_key_here'  # Same as in game_routes.py

def token_required(f):
    @wraps(f)
    def decorated(*args, **kwargs):
        token = None
        if 'Authorization' in request.headers:
            token = request.headers['Authorization'].split(" ")[1]  # Bearer <token>

        if not token:
            return jsonify({'message': 'Token is missing'}), 401

        try:
            data = jwt.decode(token, SECRET_KEY, algorithms=["HS256"])
            current_user_id = data['user_id']  # ✅ THIS MATCHES THE LOGIN TOKEN
        except jwt.ExpiredSignatureError:
            return jsonify({'message': 'Token expired'}), 401
        except jwt.InvalidTokenError:
            return jsonify({'message': 'Invalid token'}), 401

        return f(current_user_id, *args, **kwargs)  # ✅ pass user_id to route

    return decorated
