Server/requirements.txt
blinker==1.9.0
certifi==2024.7.4
charset-normalizer==3.3.2
chess==1.11.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.4
EasyProcess==1.1
entrypoint2==1.1
Flask==3.1.1
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
gTTS==2.5.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
mss==9.0.1
mysql-connector-python==9.0.0
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
plyer==2.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pypiwin32==223
pyscreenshot==3.1
python-dateutil==2.9.0.post0
pyttsx3==2.90
pytz==2024.1
pywin32==306
requests==2.32.3
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
SQLAlchemy==2.0.41
threadpoolctl==3.5.0
typing_extensions==4.14.1
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.1.3
